Create a game component
