+++ src/routes.js
@@ -2,7 +2,7 @@ const routes = require('next-routes')();
 
 routes
     .add('index', '/')
-    .add('game', '/game', 'game')
+    .add('game', '/game', 'singleplayerGame')
     .add('multiplayer_games', '/multiplayer', 'multiplayerGames')
     .add('multiplayer_game', '/multiplayer/game', 'multiplayerGame');
